improved the test rig, and added some debugging info
The test rig is entirely revamped so it now looks better.

Some of the internal debugging info, such as listing the
symbols generated, and how long it took, is now built
properly into parse and the Quby parser.

This info is also now shown on the test rig, along with
a new 'debug' option.