made expr more efficient
The left side of many expressions are the same.
For example with:
a + 4
a.foo()
a[3]

... the left side is always a variable. Before this checkout,
that variable is tested for each of the above rules. Expressions
have been re-arranged, so now it is only tested the once,
and the rest is then tacked on.

tl;dr; it does less work.