recursive changes, and moved name'd terminals into a Terminal
There is a problem with recursive rules.

So the recursive tracking code has been altered, but is
still not finished. It still includes some debugging code
for helping to work this out.

ParserRules no longer allow left-edge recursion. This is
where the left-edge of the grammar tree for a rule, cannot
lead to 'this' rule. An example of a left-edge recursive rule
is:

expr = expr plus expr

The left edge recursively leads to 'expr'. This change makes
the parser *much* more efficient, and fixes some issues with
recursive rules.

Terminals now have names built into them, and worked
out as they are defined. The previous identification code
has then been removed, to help simplify code.