removed a load of old code from the parser
The parser is moving from a tree form, to just long chains
for each statement. This allows the parser to be far more
flexible, and more efficient (as matching is more linear
and predictable).

The latter is especially important when the parser is
essentially a brute-force with indexing, and so it
dramatically reduces the search space to become
almost linear.