Changed quby.syntax to a new class structure
Originally the classes in quby.syntax just added methods
to themselves directly, in all of the constructors. This
meant you had the cost of that each time the constructor
is run.

These have all been changed to the correct way to do this;
prototypes! \o/

I have also removed a lot of the 'magic' that was going on
before, with re-arranging the methods on objects at runtime.