fixed parser bugs
The lookup table inside rules was failing to take
into account optional rules. It essentially thought
that optional rules were not optional, not it indexes
past optional ones.

A bug with array type terminals failing to pass id
and on match information to their children, was
fixed. This was an issue as the array type terminals
are removed, and only the children are folded into
the actual symbolisation section of the parser.