fixed bug with constructors being recognised as methods
Constructors were being recognised as being a method
with the name 'new'. This is now fixed, so they are seen
properly as constructors.

The issue was that the symbols used to tell if it was a
NEW keyword terminal or not, had been wrapped,
and so accessing the inner terminal was failing. It is
no longer being wrapped.