minor additions
quby.core and quby.main, with how they
interact with the quby.main.ParserInstance,
has been re-written so it will be 'locked' when parsing begins.
This makes it a little cleaner.

parse.ts now has an added 'repeat' option,
which will repeat the rule given, 0 or more times,
to create a list of rules. Just like repeatSeperator,
but without the seperator.

Altered the SyntaxList so it can be given a set of
statements when created. This makes the API a
tad cleaner, and removes unneeded empty arrays.

Added a stub for 'case when' in quby.parser.