fixed #6 operators being double generated.
When Quby was ported from JS to TypeScript,
somewhere along the way the re-balancing system
broke for operators.

Previous it worked by cloning operators, so it could
leave the original operators in place, and not move
them. This was an overly complex solution, which
wasn't actually needed.

A simpler solution, which just swaps the nodes as
needed between operators, has been implemented
to replace it.