fixed inline/preInline, and altered 'match' property on symbols
inline and preInline code is now parsed and injected correctly.

To get this working, the way retrieving the match string on Symbols,
was changed. It is now more lazy than before, and the string is only
grabbed on demand, so you now use a 'getMatch' function to grab the
match.

Extra functions like substring, and chomp, were also added to Symbol and
SourceLines,
to allow AST nodes to get a substring of their match directly, and
skipping any intermediate
strings.