added invalid function marking, and fixed bug with blocks counted as parameters
Blocks were being counted in the number of parameters
a function would take. This is fixed.

Function declarations can also be marked as being invalid,
and when done so, invalid invocations to that function are
silenced. This is to avoid receiving a huge stream of errors,
when the declaration has an issue.